Address review: rename helper, drop unused error overload, document strict mode, merge Jackson strict tests
